Remove CMake build system and migrate opengr2/libopengrn to native VS2026 projects

+++ libopengrn/crc.c
@@ -11,15 +11,14 @@ uint32_t CRC32(const uint8_t* message, size_t len)
 
     i = 0;
     crc = 0xFFFFFFFF;
-    for (; i < len; i++)
-    {
-        byte = message[i];            // Get next byte.
+    for (; i < len; i++) {
+        byte = message[i]; // Get next byte.
         crc = crc ^ byte;
-        for (j = 7; j >= 0; j--) {    // Do eight times.
-            mask = -(crc & 1);
+        for (j = 7; j >= 0; j--) { // Do eight times.
+            mask = -(int)(crc & 1);
             crc = (crc >> 1) ^ (0xEDB88320 & mask);
         }
     }
-    
+
     return ~crc;
 }
